feat: use float16 instead of bfloat16 for storage
float16 has 8x less quantization error (0.002 vs 0.016 max).
Same memory cost (2 bytes). Range [-5.9, 5.4] fits f16 fine.

+++ scripts/export_features.py
@@ -62,7 +62,7 @@ from tqdm import tqdm
 # Constants
 # -----------------------------------------------------------------------------
 
-STORAGE_DTYPE = torch.bfloat16
+STORAGE_DTYPE = torch.float16
 STORAGE_BYTES = torch.tensor([], dtype=STORAGE_DTYPE).element_size()
 
 ImageFile.LOAD_TRUNCATED_IMAGES = False
